requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.1

